Enable emulation of DAA on RL side
Before:
We calculate rewards for the common chain when it grows.
We accumulate the rewards for all new blocks on the common chain if it
grows by more than one block.
We return the time elapsed on the common chain, that is (timestamp of
last new block) - (timestamp of last old block) as part of the info
field.

Now:
We also return the number of proof-of-work puzzle solutions on the
common chain between last old block (exclusive) and last new block
(inclusive) in the info field.

This enables emulation of a DAA on the RL side.