Fix Bk attack space bug.
The attacker simulates two nodes, each having partial visibility of the DAG.
Messages visible to the 'private' node must be delivered to the private
node. Before the fix, incoming messages were only delivered to the 'public'
node.

TODO. I think this bug is present in the other Attack spaces, Nakamoto and
Bk/ll, but it is not relevant there.