Fix visibility bug in SSZ attack spaces.
With the old interface, sharing vertices was instant. During the
call `share msg`, `released msg` was updated from false to true.
The attacker's public view relied on this.

With the new interface, sharing is passive. `released msg` is updated
between calls to the nodes handler. The old public view was broken and
the emulated public node tried to access invisible vertices.

I fixed this bug by recording the just shared messages in the state and
handing them to the public node at the beginning of the next handler
evaluation.