Reimplement simulator event_handler.
Deterministic appends are now handled properly. After an append delivery
of now unlocked received vertices continues recursively. With this
change all tests for the new Tailstorm implementation pass.

Two test for bk(ll)_ssz with random policies fail. I suspect this is a
bug in the attack space, not the simulator.