Run gaussian process hyper-parameter search over night
Results are not great. I didn't even learn honest, it seems. Maybe there
is a bug in my optimization. It also seems that GP gets slower over
time. Maybe it cannot handle hundreds of observations? I read that GP
can be replaced with tree-structured Parzen estimator (TPE). Apparently
it's more appropriate for multidimensional and conditional spaces.

Overall, it's maybe not so smart to code hyper param search manually.
rl_zoo3 has a script for that. It uses optuna search package. It support
TPE out of the box.