Revive Tailstorm attack space as of June '22
In June we had successful training runs against Tailstorm(ll) discount
on varying alphas. We're not able to reproduce this against the new
Tailstorm. To rule out any problems we want to reproduce the June
results with the new simulator. Thus I revive the old protocol in the
new simulator.

As mentioned at the top of tailstorm_june.ml, the diff to tailstomll.ml
is very small. Even if we keep the attack space tailstorm_june_ssz.ml
for future reference, we probably still want to get rid of the protocol
tailstorm_june.ml

WIP: To fully enable the reproducing training runs, we have to make
observation space scaling optional. Currently tailstorm_june_ssz.ml does
not rescale to interval 0-1 thus the corresponding test in
test_protocols.py fails.