Fix parallel.sh
Since the ent_coef change to train.py, models are saved to directories
containing '.'. E.g., *-entcoef0.01-*. My regexp was not prepared for
this.

During fixing this, I also found out that some stuff is logged to `<dir>_2`
because `<dir>` already exists and is a directory. The script should now
scan both directories for model.zip files.