Hotfix install of gym<0.22 dependency.
Problem description: https://github.com/openai/gym/issues/3202

It would be much better to upgrade to a recent version of gym, or even
better, switch to gymnasium. But stable-baselines3 has not done this
upgrade/switch yet, so I'll have to wait.

When doing the switch/upgrade, don't forget to fix all three files:
- setup.py
- Makefile (_venv rule)
- .github/workflows/main.yml