mdp. uncover bug in ethereum/sm
The issue happens during truncation. We truncate ancestors of uncles
which makes these uncles new dag roots.

I think a different algorithm is required to decide what to keep an
what to truncate. I guess we have to find the common ancestor DAG-based
not history-based when truncating the history of the common ancestor.
This would allow specifying Ethereum with head = pref, not like now with
the preferred block lagging self.horizon behind head.

Another consideration is to reward forward (in direction of children)
instead of backward (in direction of parents). But this seems to be
not intuitive.