mdp. improve state space cutoff
Before, we just disallowed Continue when the state space grew too big.
Disallowing Continue implies termination and hence no future rewards.
This may be okay for the traditional MDP, where at least one of
Match/Override/Adopt is always feasible and implies jumping back into
the restricted state space. In our case only the Adopt is feasible
(through Consider) but not the Match/Override. This is because our
Release only marks for release while the execution follows in the
next Continue actions.

The new approach introduces a forth action, Communicate, which executes
communication like Continue but does not append a new block. This forces
the attacker to release (and consider) all blocks before actually
entering a terminal state.