generic_v1. highlight state size scaling gap to v0
We can explore Bitcoin to distance 17 with less than 100k states using
generic_v0 but only to distance 8 with generic_v1.

Is this due to the more aggressive (but more specific) common chain
truncation in v0?

Another delta is that Bitcoin's state.head goes into the vertex coloring
of DAG normalization in v0 but not in v1.