generic_v1/ghostdag. fix history/reward computation
Ghostdag distinguishes blue blocks and non-blue blocks. The hypothesis
is that non-blue blocks originate from malicious behaviour (or the
weaker part of a split network). The Ghostdag paper is not very explicit
about reward allocation, but the Kaspa documentation says that only blue
blocks are assigned a reward.

According to the paper, Ghostdag's goal is to establish an order where
blue blocks take precedence over non-blue blocks. Kaspa, the algorithm
in the paper, and the spec version before this fix include non-blue
blocks in the history. This allows merging non-conflicting transactions
mentioned in non-blue blocks.

These details do not matter here, as we use the history only for
allocating rewards. With this fix, I ensure that only blue blocks end up
in the history and hence only blue blocks allocate a reward.