generic_v1. revise common chain truncation
The problem here was: the heuristic allowed for partially removing
parents of blocks. This breaks the semantics of the BlockDAG and
is not safe.

The new version should avoid this case; we now check that none of the
removed blocks has children among the future of the new genesis block.